requirements.txt
streamlit==1.41.0
groq==0.25.0
PyPDF2==3.0.1
pandas==2.2.3
python-dotenv==1.0.1
fpdf2==2.8.2
